refactor: move classes into single source file
