+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/datasource/DataSourceImpl.kt
@@ -57,9 +57,3 @@ internal abstract class DataSourceImpl<T>(
         }
     }
 }
-
-/**
- * Syntactic sugar for when no input validation is required. Developers must explicitly state
- * this is the case.
- */
-internal val NoInputValidation = { true }
