+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/datasource/NoInputValidation.kt
@@ -1,0 +1,7 @@
+package io.embrace.android.embracesdk.arch.datasource
+
+/**
+ * Syntactic sugar for when no input validation is required. Developers must explicitly state
+ * this is the case.
+ */
+internal val NoInputValidation = { true }
