+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/arch/schema/EmbraceAttributeKey.kt
@@ -1,0 +1,25 @@
+package io.embrace.android.embracesdk.arch.schema
+
+import io.embrace.android.embracesdk.internal.spans.toEmbraceAttributeName
+import io.opentelemetry.api.common.AttributeKey
+
+/**
+ * Defines a unique object to represent a [EmbraceAttribute]
+ */
+internal class EmbraceAttributeKey(
+    override val id: String,
+    otelAttributeKey: AttributeKey<String>? = null,
+    useIdAsAttributeName: Boolean = false,
+    isPrivate: Boolean = false
+) : EmbraceAttribute {
+    override val name: String = if (!useIdAsAttributeName && otelAttributeKey?.key != null) {
+        otelAttributeKey.key
+    } else {
+        id.toEmbraceAttributeName(isPrivate)
+    }
+
+    /**
+     * The [AttributeKey] equivalent for this object to be used in conjunction with OTel objects
+     */
+    val attributeKey: AttributeKey<String> = otelAttributeKey ?: AttributeKey.stringKey(name)
+}
